move sizing defaults to k8sutils
