+++ cli/cmd/resources/autoscaler.go
@@ -9,6 +9,7 @@ import (
 	"github.com/odigos-io/odigos/cli/pkg/kube"
 	"github.com/odigos-io/odigos/common"
 	"github.com/odigos-io/odigos/common/consts"
+	"github.com/odigos-io/odigos/k8sutils/pkg/sizing"
 
 	admissionregistrationv1 "k8s.io/api/admissionregistration/v1"
 	appsv1 "k8s.io/api/apps/v1"
@@ -545,9 +546,9 @@ func (a *autoScalerResourceManager) InstallFromScratch(ctx context.Context) erro
 	var resourceReqs corev1.ResourceRequirements
 
 	if a.config.Autoscaler == nil {
-		resourceReqs = GetDefaultResourceRequirements()
+		resourceReqs = sizing.GetDefaultResourceRequirements()
 	} else {
-		resourceReqs = GetResourceRequirementsWithDefaults(a.config.Autoscaler.ResourceConfig)
+		resourceReqs = sizing.GetResourceRequirementsWithDefaults(a.config.Autoscaler.ResourceConfig)
 	}
 
 	resources := []kube.Object{
