+++ cli/cmd/resources/instrumentor.go
@@ -10,6 +10,7 @@ import (
 	"github.com/odigos-io/odigos/cli/pkg/kube"
 	"github.com/odigos-io/odigos/common"
 	"github.com/odigos-io/odigos/common/consts"
+	"github.com/odigos-io/odigos/k8sutils/pkg/sizing"
 
 	admissionregistrationv1 "k8s.io/api/admissionregistration/v1"
 	appsv1 "k8s.io/api/apps/v1"
@@ -703,9 +704,9 @@ func (a *instrumentorResourceManager) InstallFromScratch(ctx context.Context) er
 	var resourceReqs corev1.ResourceRequirements
 
 	if a.config.Instrumentor == nil {
-		resourceReqs = GetDefaultResourceRequirements()
+		resourceReqs = sizing.GetDefaultResourceRequirements()
 	} else {
-		resourceReqs = GetResourceRequirementsWithDefaults(a.config.Instrumentor.ResourceConfig)
+		resourceReqs = sizing.GetResourceRequirementsWithDefaults(a.config.Instrumentor.ResourceConfig)
 	}
 
 	resources := []kube.Object{
