+++ cli/cmd/resources/scheduler.go
@@ -9,6 +9,7 @@ import (
 	"github.com/odigos-io/odigos/cli/pkg/kube"
 	"github.com/odigos-io/odigos/common"
 	"github.com/odigos-io/odigos/common/consts"
+	"github.com/odigos-io/odigos/k8sutils/pkg/sizing"
 
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -335,9 +336,9 @@ func (a *schedulerResourceManager) InstallFromScratch(ctx context.Context) error
 	var resourceReqs corev1.ResourceRequirements
 
 	if a.config.Scheduler == nil {
-		resourceReqs = GetDefaultResourceRequirements()
+		resourceReqs = sizing.GetDefaultResourceRequirements()
 	} else {
-		resourceReqs = GetResourceRequirementsWithDefaults(a.config.Scheduler.ResourceConfig)
+		resourceReqs = sizing.GetResourceRequirementsWithDefaults(a.config.Scheduler.ResourceConfig)
 	}
 
 	resources := []kube.Object{
